add and edit user up and running
