+++ client/src/app/admin/admin-routing.module.ts
@@ -5,10 +5,12 @@ import { CompanyComponent } from "./company/company.component";
 import { UserComponent } from "./user/user.component";
 import { AddUserComponent } from "./user/adduser/add-user.component";
 import { EditUserComponent } from "./user/edituser/edit-user.component";
+import {StatisticComponent} from "./statistic/statistic.component";
 
 const routes: Routes = [
   {path: 'admin', component: AdminComponent, children: [
       { path: 'company', component: CompanyComponent },
+      { path: 'statistic', component: StatisticComponent },
       { path: 'user', component: UserComponent, children: [
           { path: 'adduser', component: AddUserComponent },
           { path: 'edituser', component: EditUserComponent }
