+++ client/src/app/admin/statistic/statistic.component.html
@@ -1,0 +1,3 @@
+<p>
+  statistic works!
+</p>
