+++ client/src/app/admin/statistic/statistic.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-statistic',
+  templateUrl: './statistic.component.html',
+  styleUrls: ['./statistic.component.css']
+})
+export class StatisticComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
