+++ client/src/app/admin/user/adduser/add-user.component.css
@@ -1,0 +1,10 @@
+mat-card {
+  display: flex;
+  flex-direction: column;
+  width: 30%;
+  margin: 0 auto;
+}
+
+mat-radio-button {
+  margin: 0 10px 5px 0;
+}
