+++ client/src/app/admin/user/user.component.html
@@ -1,61 +1,10 @@
 <!--
-<mat-drawer-container class="container">
-  <mat-drawer mode="side" opened class="dashboard" #drawer>
-    <mat-radio-button color="primary" (change)="dashboardSelection(1)">Show all users</mat-radio-button>
-    <mat-radio-button color="primary" (change)="dashboardSelection(2)">Add user</mat-radio-button>
-    <mat-radio-button color="primary" (change)="dashboardSelection(3)">Search User</mat-radio-button>
-  </mat-drawer>
 
   <mat-drawer-content>
 
     <ng-container *ngIf="!choice; else showUsers"></ng-container>
     <ng-container *ngIf="!addUser; else addAdmin"></ng-container>
     <ng-container *ngIf="!showStatisticOnUser; else showUser"></ng-container>
-
-
-    <ng-template #showUser>
-      <mat-card [formGroup]="showUserForm">
-        <mat-card-header>Search user</mat-card-header>
-
-          <div class="searchFieldAndIcon inputIconBg">
-            <input type="text" placeholder="Search user" [matAutocomplete]="auto" [formControl]="search">
-            <i class="fa fa-search" aria-hidden="true"></i>
-            <mat-autocomplete #auto="matAutocomplete" (optionSelected)="userSearch($event.option.value)">
-              <mat-option *ngFor="let user of usersFound" [value]="user.user">{{user.user}}</mat-option>
-            </mat-autocomplete>
-          </div>
-      </mat-card>
-    </ng-template>
-
-    <ng-template #addAdmin>
-      <mat-card [formGroup]="userForm" novalidate>
-        <mat-card-header>Add a user</mat-card-header>
-
-        <mat-form-field>
-          <input id="user" formControlName="user" matInput placeholder="Name" />
-          <mat-error *ngIf="(user.dirty || user.touched) && user.invalid">Name required</mat-error>
-        </mat-form-field>
-
-        <mat-form-field>
-          <input id="email" matInput placeholder="Email" formControlName="email" class="form-control" type="email" required>
-          <mat-error *ngIf="(email.dirty || email.touched) && email.invalid">Email required</mat-error>
-        </mat-form-field>
-
-        <mat-form-field>
-          <input id="password" matInput placeholder="Password" type="password" formControlName="password" class="form-control" required >
-          <mat-error *ngIf="(password.dirty || password.touched) && password.invalid">Password required</mat-error>
-        </mat-form-field>
-
-        <mat-radio-group formControlName="role">
-          <mat-radio-button name="role"  color="primary" value="1">User</mat-radio-button>
-          <mat-radio-button name="role"  color="primary" value="2">Admin</mat-radio-button>
-        </mat-radio-group>
-
-        <mat-card-actions>
-          <button mat-raised-button type="button" [disabled]="!userForm.valid" (click)="onSubmit()" color="primary">Add User</button>
-        </mat-card-actions>
-      </mat-card>
-    </ng-template>
 
     <ng-template #showUsers>
       <mat-form-field>
